change all source quantities back to floats

+++ include/openmc/random_ray/random_ray.h
@@ -93,7 +93,7 @@ private:
 
   //---------------------------------------------------------------------------
   // Private data members for kinetic simulations
-  vector<double> delta_psi_prime_;
+  vector<float> delta_psi_prime_;
 
 }; // class RandomRay
